requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
pgvector>=0.2,<1.0
dj-database-url>=2.1,<3.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-results>=2.5,<3.0

# AI/ML
langchain>=0.1,<1.0
langchain-openai>=0.0.5,<1.0
langchain-community>=0.0.20,<1.0
openai>=1.10,<2.0
tiktoken>=0.5,<1.0

# Utilities
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
structlog>=24.1,<25.0

# Development
ipython>=8.0,<9.0

